Sqlite: Add range getter
Sometimes it is handy to return a range so you can break in a for range
loop. So you can now write:

 for (auto [key, value] :
      statement.range<std::tuple<long long, std::string>>()) {
    values.push_back(value);
    if (value == Tuple{"foo", 23.3, 2})
         break;
}

And it will only step over the statement to the break. So you don't fetch
not used values anymore.

But be careful. The range is view to the statement. Holding it longer or
reusing it can lead to strange behavior because the state of the
statement is changing.

Change-Id: I90191f7da5a015c7d8077d5bc428655a2b53246c